refactor: organize supervised model module
