+++ README.md
@@ -2,16 +2,16 @@
 [![Crates.io](https://img.shields.io/crates/v/automl.svg)](https://crates.io/crates/automl)
 [![docs.rs](https://img.shields.io/docsrs/automl/latest?logo=rust)](https://docs.rs/automl)
 
-# AutoML with SmartCore
+# `AutoML` with `SmartCore`
 
-AutoML (_Automated Machine Learning_) streamlines machine learning workflows, making them more accessible and efficient
+`AutoML` (_Automated Machine Learning_) streamlines machine learning workflows, making them more accessible and efficient
 for users of all experience levels. This crate extends the [`smartcore`](https://docs.rs/smartcore/) machine learning
 framework, providing utilities to
 quickly train, compare, and deploy models.
 
 # Install
 
-Add AutoML to your `Cargo.toml` to get started:
+Add `AutoML` to your `Cargo.toml` to get started:
 
 **Stable Version**
 
@@ -27,12 +27,12 @@ automl = { git = "https://github.com/cmccomb/rust-automl" }
 
 # Example Usage
 
-Here’s a quick example to illustrate how AutoML can simplify model training and comparison:
+Here’s a quick example to illustrate how `AutoML` can simplify model training and comparison:
 
 ```rust
 let (x, y) = automl::regression_testing_data();
 let settings = automl::Settings::default_regression();
-let mut regressor = automl::SupervisedModel::new(x, y, settings);
+let mut regressor = automl::supervised_model::SupervisedModel::new(x, y, settings);
 regressor.train();
 ```
